refactor(v1): turn utils into a package; group helpers by concern
v1/utils.py and cli/log.py held unrelated helpers in flat modules. Make verifiers/v1/utils a
package split by concern: format (display), install (env-id / on-demand hub install), logging
(was cli/log.py — used by every CLI, not just eval), memory (worker RSS trimming). Consumers
import from the specific submodule.
