+++ verifiers/v1/cli/dashboard/eval.py
@@ -23,7 +23,7 @@ from verifiers.v1.cli.output import output_path
 from verifiers.v1.configs.eval import EvalConfig
 from verifiers.v1.rollout import Phase, Rollout
 from verifiers.v1.trace import Trace
-from verifiers.v1.utils import format_count, format_reward, format_time
+from verifiers.v1.utils.format import format_count, format_reward, format_time
 
 # For sizing pages to the terminal: detects the real terminal height/width each access (the live
 # view writes to the same terminal). Reused so we don't rebuild it every refresh tick.
