+++ verifiers/v1/cli/eval/runner.py
@@ -103,7 +103,7 @@ async def run_eval_server(config: EvalConfig) -> list[Trace]:
     import multiprocessing as mp
     from functools import partial
 
-    from verifiers.v1.cli.log import setup_logging
+    from verifiers.v1.utils.logging import setup_logging
     from verifiers.v1.env import pool_serve_kwargs
     from verifiers.v1.serve import EnvClient, env_config_data, serve_env
