+++ verifiers/v1/utils/__init__.py
@@ -1,0 +1,3 @@
+"""Small shared helpers for v1, by concern: `format` (display), `install` (env-id /
+on-demand hub install), `logging` (route stdlib logs through loguru), `memory` (worker RSS).
+Import from the submodule, e.g. `from verifiers.v1.utils.format import format_reward`."""
